LayoutTests/css3/scroll-snap/scroll-snap-initial-expected.txt

Tests 'initial' on scroll snap properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS noInitial.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('scroll-snap-type') is "both proximity"
PASS noInitial.style['scroll-padding'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('scroll-padding') is "20px"
PASS noInitial.style['scroll-snap-margin'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('scroll-snap-margin') is "1px"
PASS noInitial.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('scroll-snap-align') is "center center"
PASS initialType.style['scroll-snap-type'] is "initial"
PASS window.getComputedStyle(initialType).getPropertyValue('scroll-snap-type') is "none"
PASS initialType.style['scroll-padding'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('scroll-padding') is "20px"
PASS initialType.style['scroll-snap-margin'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('scroll-snap-margin') is "1px"
PASS initialType.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('scroll-snap-align') is "center center"
PASS initialPadding.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(initialPadding).getPropertyValue('scroll-snap-type') is "both proximity"
PASS initialPadding.style['scroll-padding'] is "initial"
PASS window.getComputedStyle(initialPadding).getPropertyValue('scroll-padding') is "0px"
PASS initialPadding.style['scroll-snap-margin'] is ""
PASS window.getComputedStyle(initialPadding).getPropertyValue('scroll-snap-margin') is "1px"
PASS initialPadding.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(initialPadding).getPropertyValue('scroll-snap-align') is "center center"
PASS initialMargin.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(initialMargin).getPropertyValue('scroll-snap-type') is "both proximity"
PASS initialMargin.style['scroll-padding'] is ""
PASS window.getComputedStyle(initialMargin).getPropertyValue('scroll-padding') is "20px"
PASS initialMargin.style['scroll-snap-margin'] is "initial"
PASS window.getComputedStyle(initialMargin).getPropertyValue('scroll-snap-margin') is "0px"
PASS initialMargin.style['scroll-snap-align'] is ""
PASS window.getComputedStyle(initialMargin).getPropertyValue('scroll-snap-align') is "center center"
PASS initialAlign.style['scroll-snap-type'] is ""
PASS window.getComputedStyle(initialAlign).getPropertyValue('scroll-snap-type') is "both proximity"
PASS initialAlign.style['scroll-padding'] is ""
PASS window.getComputedStyle(initialAlign).getPropertyValue('scroll-padding') is "20px"
PASS initialAlign.style['scroll-snap-margin'] is ""
PASS window.getComputedStyle(initialAlign).getPropertyValue('scroll-snap-margin') is "1px"
PASS initialAlign.style['scroll-snap-align'] is "initial"
PASS window.getComputedStyle(initialAlign).getPropertyValue('scroll-snap-align') is "none none"
PASS allInitial.style['scroll-snap-type'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('scroll-snap-type') is "none"
PASS allInitial.style['scroll-padding'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('scroll-padding') is "0px"
PASS allInitial.style['scroll-snap-margin'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('scroll-snap-margin') is "0px"
PASS allInitial.style['scroll-snap-align'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('scroll-snap-align') is "none none"
PASS successfullyParsed is true

TEST COMPLETE
